requirements.txt
pymupdf
easyocr
pandas
googletrans==4.0.0-rc1
langchain 
langchain_community
langchain_google_genai
pdf2image
numpy
python-docx
faiss-cpu
werkzeug
sentence-transformers
PyPDF2
pypdf
accelerate>=0.26.0
eventlet  # Important for Flask-SocketIO
python-dotenv
spacy
pdfplumber
pillow
openpyxl
uuid

torch
torchvision

flask
flask-socketio


